Implement rest layer for Item

+++ src/main/java/io/github/bluething/myboostposystem/rest/ItemResponse.java
@@ -3,12 +3,12 @@ package io.github.bluething.myboostposystem.rest;
 import java.time.LocalDateTime;
 
 record ItemResponse(Integer id,
-                           String name,
-                           String description,
-                           Integer price,
-                           Integer cost,
-                           String createdBy,
-                           String updatedBy,
-                           LocalDateTime createdDatetime,
-                           LocalDateTime updatedDatetime) {
+                    String name,
+                    String description,
+                    Long price,
+                    Long cost,
+                    String createdBy,
+                    String updatedBy,
+                    LocalDateTime createdDatetime,
+                    LocalDateTime updatedDatetime) {
 }
